# packages/contracts/openapi/bridge-api.yaml
openapi: 3.0.3
info:
  title: LR-FrameFlow Bridge ↔ API
  version: "0.1.0"
  description: |
    Thin Gateway: Jobs werden asynchron eingereiht (`202 Accepted`).
    Zusätzliche kanonische JSON Schemas unter `json-schema/` (gleiche Felder abstimmen).
servers:
  - url: http://localhost:8080
    description: Local API (compose default host port)
paths:
  /health:
    get:
      operationId: healthCheck
      summary: Liveness
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
  /v1/photos:
    post:
      operationId: uploadPhoto
      summary: Upload JPEG preview + metadata, returns photo_id
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, metadata]
              properties:
                file:
                  type: string
                  format: binary
                  description: JPEG preview image (max longest side 1200px recommended)
                metadata:
                  type: string
                  description: JSON string conforming to photo-upload-v1 schema
      responses:
        "200":
          description: Photo registered (already existed or newly created)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PhotoCreatedResponse"
  /v1/profiles:
    get:
      operationId: listProfiles
      summary: List all profiles
      responses:
        "200":
          description: List of profiles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProfileResponse"
  /v1/profiles/{profile_id}:
    get:
      operationId: getProfile
      summary: Get a single profile by ID
      parameters:
        - name: profile_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Profile found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfileResponse"
        "404":
          description: Profile not found
  /v1/jobs/train:
    post:
      operationId: createTrainJob
      summary: Enqueue profile training job
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TrainJobRequestV1"
      responses:
        "202":
          description: Job queued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobAcceptedResponse"
  /v1/jobs/edit:
    post:
      operationId: createEditJob
      summary: Enqueue edit inference job
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EditJobRequestV1"
      responses:
        "202":
          description: Job queued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobAcceptedResponse"
  /v1/jobs/{job_id}:
    get:
      operationId: getJob
      summary: Retrieve job lifecycle state
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Known job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobSnapshotV1"
        "404":
          description: Job not found
  /v1/jobs/{job_id}/result:
    get:
      operationId: getJobResult
      summary: Retrieve edit result (LR settings) once job is completed
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Edit result available
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EditResultResponse"
        "404":
          description: Job or result not found
        "409":
          description: Job not yet completed
  /v1/jobs/{job_id}/events:
    get:
      operationId: streamJobEvents
      summary: Server-Sent Events stream (optional MVP)
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: SSE stream placeholder
          content:
            text/event-stream:
              schema:
                type: string
components:
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema:
        type: string
        description: Replay-safe enqueue key for Bridge retries
  schemas:
    JobAcceptedResponse:
      type: object
      required: [job_id, status, kind]
      properties:
        job_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [queued]
        kind:
          type: string
          enum: [train, edit]
    TrainJobRequestV1:
      type: object
      required:
        - schema_version
        - profile_name
        - genre
        - format_type
        - color_type
        - photo_ids
      properties:
        schema_version:
          type: string
          example: train-request-v1
        correlation_id:
          type: string
          description: Trace id from Lightroom/Bridge pipeline
        profile_name:
          type: string
        genre:
          type: string
        format_type:
          type: string
          enum: [raw, jpeg_tiff]
        color_type:
          type: string
          enum: [color, bw]
        photo_ids:
          type: array
          minItems: 1
          items:
            type: string
            format: uuid
        base_preset:
          type: object
          additionalProperties: true
    EditJobRequestV1:
      type: object
      required:
        - schema_version
        - profile_id
        - photo_ids
      properties:
        schema_version:
          type: string
          example: edit-request-v1
        correlation_id:
          type: string
        profile_id:
          type: string
          format: uuid
        collection_name:
          type: string
        photo_ids:
          type: array
          minItems: 1
          items:
            type: string
            format: uuid
    JobSnapshotV1:
      type: object
      properties:
        job_id:
          type: string
          format: uuid
        status:
          type: string
          enum:
            - queued
            - running
            - completed
            - failed
        payload_schema_version:
          type: string
        correlation_id:
          type: string
    PhotoCreatedResponse:
      type: object
      required: [photo_id, s3_key, already_existed]
      properties:
        photo_id:
          type: string
          format: uuid
        s3_key:
          type: string
        already_existed:
          type: boolean
    ProfileResponse:
      type: object
      required: [profile_id, name, genre, format_type, color_type, status, version]
      properties:
        profile_id:
          type: string
          format: uuid
        name:
          type: string
        genre:
          type: string
        format_type:
          type: string
          enum: [raw, jpeg_tiff]
        color_type:
          type: string
          enum: [color, bw]
        status:
          type: string
          enum: [draft, training, ready, failed]
        version:
          type: integer
        model_artifact_key:
          type: string
          nullable: true
        failure_reason:
          type: string
          nullable: true
    EditResultResponse:
      type: object
      required: [job_id, photo_id, lr_settings]
      properties:
        job_id:
          type: string
          format: uuid
        photo_id:
          type: string
          format: uuid
        lr_settings:
          type: object
          additionalProperties: true
          description: Lightroom Develop settings to apply (exposure, contrast, highlights, etc.)
        s3_key:
          type: string
          nullable: true
    JobEnvelopeV1:
      type: object
      description: Redis queue payload (JSON Schema job-envelope-v1.schema.json)
      required:
        - envelope_schema_version
        - job_id
        - job_kind
        - payload
      additionalProperties: false
      properties:
        envelope_schema_version:
          type: string
          enum: ["job-envelope-v1"]
        job_id:
          type: string
          format: uuid
        job_kind:
          type: string
          enum: [train, edit]
        attempt:
          type: integer
          minimum: 0
          default: 0
        trace_context:
          type: object
          additionalProperties: {}
        payload:
          type: object
          additionalProperties: {}